+++ CHANGES.rst
@@ -2,6 +2,11 @@
 CHANGES
 =======
 
+current
+-------
+- refactor tests to pytest
+- weaker pyramid_yml requirements. Use ``registry['config']`` instead of ``request.config`` which gets added only when explicitly including tzf.pyramid_yml package.
+
 0.1.0 (prerelease)
 ------------------
 - license information